+++ src/VortexPasta.jl
@@ -12,7 +12,9 @@ using .Quadratures
 export GaussLegendre
 
 include("Filaments/Filaments.jl")
+include("FindNearbySegments/FindNearbySegments.jl")
 include("FilamentIO/FilamentIO.jl")
+
 include("Reconnections/Reconnections.jl")
 
 include("BiotSavart/BiotSavart.jl")