+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Move reconnections to separate `Reconnections` module.
 
+- Move most of the short-range backend implementation to new `FindNearbySegments` module.
+
 ## [0.4.0] - 2023-10-16
 
 ### Added